app/requirements.txt
fastapi==0.115.7
uvicorn[standard]==0.23.2
python-dotenv~=1.0.1
scikit-learn==1.3.2
schedule==1.2.2
requests==2.31.0
numpy==1.24.4
starlette==0.45.3
httpx==0.28.1
pytest==8.3.4
